[Starbucks Kiosk]  네비게이션 콘텐츠 코드 분할
